+++ player/src/foc/foc_interp.h
@@ -1,6 +1,6 @@
 #ifndef FOC_INTERP_H
 #define FOC_INTERP_H
 
-void foc_interp_init(std::vector<double>*, int, int, float);
-bool foc_interp_update(float*, std::vector<double>*);
+void foc_interp_init(std::vector<float>*, int, int, float);
+bool foc_interp_update(float*, std::vector<float>*);
 #endif
